LayoutTests/imported/w3c/web-platform-tests/2dcontext/path-objects/2d.path.isPointInPath.basic.1-expected.txt

2d.path.isPointInPath.basic.1
isPointInPath() detects whether the point is inside the path
Actual output:

PASS isPointInPath() detects whether the point is inside the path 
